fix: port FlareSolverr detector fallback and async PlaylistExport handlers
- FlareDetector (Tubifarry a8ab7d7): challenges behind proxies without a
  Cloudflare Server header are caught via body inspection; non-CF servers
  with no definitive indicator no longer count, reducing false positives.
- FlareCookie.expires is double — FlareSolverr returns fractional epoch
  values that previously failed deserialization and killed cookie handling.
- PlaylistExport (Tubifarry 3d9adc4): schema-refresh handlers moved to
  IHandleAsync so a slow provider walk cannot block the synchronous event
  bus during startup.

+++ src/Sleezer/Notifications/FlareSolverr/FlareRecords.cs
@@ -11,7 +11,7 @@ public record FlareCookie(
     [property: JsonPropertyName("value")] string Value,
     [property: JsonPropertyName("domain")] string Domain,
     [property: JsonPropertyName("path")] string Path = "/",
-    [property: JsonPropertyName("expires")] int Expiry = 0,
+    [property: JsonPropertyName("expires")] double Expiry = 0,
     [property: JsonPropertyName("httpOnly")] bool HttpOnly = false,
     [property: JsonPropertyName("secure")] bool Secure = false,
     [property: JsonPropertyName("sameSite")] string? SameSite = null)
